+++ phantom/dev/ia32/uos_i8042.c
@@ -6,25 +6,14 @@
 
 #include <ia32/pio.h>
 #include <phantom_libc.h>
-
 #include <kernel/device.h>
 #include <kernel/drivers.h>
 #include <hal.h>
 #include <errno.h>
-
-//#include <video/screen.h>
-
-//#include <event.h>
-//#include <queue.h>
-
 #include <threads.h>
-//#include <kernel/config.h>
-//#include <time.h>
 
 #include <compat/uos.h>
-
-
-#include "uos_i8042.h"
+#include <ia32/pc/uos_i8042.h>
 
 /*
  * Clear input/output buffers of keyboard controller.
