+++ phantom/libkern/makefile
@@ -4,7 +4,7 @@ BUILD_ROOT = $(realpath $(PHANTOM_HOME))/phantom
 
 TARGET	= libkern.a
 
-SRCDIRS += $(BUILD_ROOT)/libkern $(BUILD_ROOT)/libkern/$(ARCH)
+SRCDIRS += $(BUILD_ROOT)/libkern $(BUILD_ROOT)/libkern/snmp $(BUILD_ROOT)/libkern/$(ARCH)
 #SRCDIRS += $(BUILD_ROOT)/libkern/net
 
 ifeq ($(ARCH),ia32)
